🏗️ Restructured login component path
