+++ src/pages/login/component/Login.jsx
@@ -1,4 +1,3 @@
-import { logDOM, render } from '@testing-library/react';
 import React from 'react';
 import styles from "./Login.module.scss";
